fix(connector): carry ids as 64-bit EntityId on Unity 6000.6
Unity 6000.6 escalated the int <-> EntityId implicit conversions from warning
(CS0618) to obsolete-as-error (CS0619), so EntityIdCompat no longer compiles; and
at runtime both implicit operators - like EditorUtility.InstanceIDToObject(int) -
now throw System.NotImplementedException. Entity ids are 64-bit there
(EntityId.ToULong / FromULong / Object.GetEntityId / EditorUtility.EntityIdToObject),
EntityIdToObject(FromULong(low32Only)) does not resolve, and
FindObjectFromInstanceID now takes an EntityId, so a bare 32-bit legacy id can no
longer resolve an object at all and the int instance_id contract cannot be honoured.

Carry ids as ulong instead: EntityIdCompat.IdOf returns ulong, ToObject takes ulong,
with the version gate kept intact
  >= 6000.6  EntityId.ToULong / EntityId.FromULong
  6000.3-5   the previous reflected-operator path, unchanged
  < 6000.3   legacy int, bit-preserved ((ulong)(uint) / (int)(uint))
ToolParams gains GetULong; tool parameter declarations and JSON parsers move from
int to ulong; the affected test sites follow.

Verified on Unity 6000.6.2f1 in a live editor session: Tundra build success and
HeraAgent.Editor.dll emitted, `status` ready, four domain reloads survived with no
human intervention and a stable port, and an EventSystem-driven click on the main
menu reached the difficulty screen. The 6000.3/6000.5 branch is unchanged by this
commit but was not exercised here.

+++ AgentConnector/Editor/Core/InputQaTypes.cs
@@ -53,7 +53,7 @@ namespace HeraAgent
     internal sealed class InputQaHit
     {
         public int rank;
-        public int instance_id;
+        public ulong instance_id;
         public string name;
         public string path;
         public string module;
@@ -66,7 +66,7 @@ namespace HeraAgent
     {
         public InputQaOptions Options;
         public EventSystem EventSystem;
-        public int TargetId;
+        public ulong TargetId;
         public string TargetName;
         public string TargetPath;
         public Vector2 Point;
